Moved utils.ts to componenents/Icons/index.ts
